% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cachemean}
\alias{cachemean}
\title{Mean of Vector with cache}
\usage{
cachemean(x, ...)
}
\arguments{
\item{x}{vector to which the mean function will be calculated for}

\item{...}{additional arguments to the regular \code{mean} function}
}
\description{
\code{cachemean} Computes the mean of a vector only if the value is not in cache
}
\details{
The method intelligently uses cached data to avoid re-calculation
}
\examples{
x <- makeVector(c(10,20,5,18))
cachemean(x)
}
\author{
Mario Gamboa
}
\keyword{cache}

