% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
\title{Computes the inverse matrix.}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{makeCacheMatrix}
}
\value{
Inverse matrix.
}
\description{
\code{makeCacheMatrix} Computes the inverse matrix.
}
\details{
The function uses the function makeCacheMatrix to store the matrix to resolve
If makeCacheMAtrix has the computation stored, returns the stored value
If not, this function computes it and stores in makeCacheMatrix cache
}

