% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Creates an object that can store the computed inverse matrix.}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{Matrix}
}
\value{
makeCacheMatrix "object". function with internal data and other functions
}
\description{
\code{makeCacheMatrix} Creates an object that can store the computed inverse matrix
}
\details{
The function creates an object with 4 "methods" that could be used to compute
and store an inverse matrix.
Aquesta funció afegeix zeros a l'esquerra d'una cadena fins que aquesta té
el nombre indicat de zeros a l'esquerra
}
\examples{
matriu <- matrix( c(2,1,1,2), 2,2)
Cmatriu <- makeCacheMatrix(matriu)
inversa <- cacheSolve(Cmatriu)
}

