% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeVector}
\alias{makeVector}
\title{\code{makeVector} creates a special "vector", which is a list containing a functions to:
* set the value of the vector
* get the value of the vector
* set the value of the mean
* get the value of the mean}
\usage{
makeVector(x = numeric())
}
\arguments{
\item{x}{the vector to which the mean function will be calculated for}
}
\value{
a list of functions associated with the vector
}
\description{
\code{makeVector} creates a special "vector", which is a list containing a functions to:
* set the value of the vector
* get the value of the vector
* set the value of the mean
* get the value of the mean
}

