% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{MyCustomMatrix-class}
\alias{MyCustomMatrix}
\alias{MyCustomMatrix-class}
\title{Reference Class emulating the OOP model in R for this assignment.
\strong{Just for learning purposes}
Moreover, instead of using \code{identical} function for checking if the
matrix changed or not I use the \code{digest} library, using a hashing
function (md5) to cache the inverse matrix value to know it. Also, the
library \code{matrixcalc} is used to check if the matrix passed as an
argument is singular or not; in that case the execution will stop or not
(because you cannot invert a singular matrix)}
\arguments{
\item{data}{the matrix with inverse matrix caching capabilities}
}
\description{
Reference Class emulating the OOP model in R for this assignment.
\strong{Just for learning purposes}
Moreover, instead of using \code{identical} function for checking if the
matrix changed or not I use the \code{digest} library, using a hashing
function (md5) to cache the inverse matrix value to know it. Also, the
library \code{matrixcalc} is used to check if the matrix passed as an
argument is singular or not; in that case the execution will stop or not
(because you cannot invert a singular matrix)
}
\examples{
myMatrix <- matrix(rnorm(1e6), 5,5)
myObj <- MyCustomMatrix$new(data = myMatrix)
myObj$getinverse()
}
\keyword{Coursera,}
\keyword{Programming,}
\keyword{ProgrammingAssignment2}
\keyword{R}

