% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cachemean}
\alias{cachemean}
\title{\code{cachemean} function that computes the mean of a vector only if the value is not in cache}
\usage{
cachemean(x, ...)
}
\arguments{
\item{x}{the vector to which the mean function will be calculated for}
}
\value{
the mean value for the vector
}
\description{
\code{cachemean} function that computes the mean of a vector only if the value is not in cache
}

