% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cachematrix.R
\name{cacheSolve}
\alias{cacheSolve}
\title{A object that solves the retrieving of an inverse of a matrix
in a makeCacheMatrix object}
\usage{
cacheSolve(matObj, ...)
}
\arguments{
\item{matObj}{list object returned by makeCacheMatrix}

\item{...}{aditional parameter to ginv function}
}
\value{
the inverse matrix
}
\description{
A object that solves the retrieving of an inverse of a matrix
in a makeCacheMatrix object
}
\examples{
m <- cacheSolve(makeCacheMatrix(matrix(nrow=3, ncol=3, 3)))
}
\author{
Andre B.
}
\seealso{
\code{\link{makeCacheMatrix}} function for creating the matObj,
\code{\link{ginv}} for getting the inverse of matrix
}

