% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
\title{Check if there is an inverse matrix cached. If so, it calculates the
inverse matrix.}
\usage{
cacheSolve(customMatrix)
}
\arguments{
\item{customMatrix}{the customMatrix object generated in makeCacheMatrix}
}
\description{
\code{customMatrix} returns a custom matrix with the inverse matrix cached
}
\examples{
cacheSolve(customMatrix)
}
\keyword{Coursera,}
\keyword{Programming,}
\keyword{ProgrammingAssignment2}
\keyword{R}

