% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
\title{Calculates the inverse of a matrix}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{special matrix object with cache capabilities}

\item{...}{additional arguments to the regular \code{solve} function}
}
\description{
\code{cacheSolve} computes the inverse of a matrix
}
\details{
The method intelligently uses cached data to avoid re-calculation
}
\examples{
x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))
cacheSolve(x)
}
\author{
Mario Gamboa
}
\keyword{cache}

