% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
<<<<<<< HEAD
\title{Calculates the inverse of a matrix}
=======
\title{\code{cacheSolve} function that computes the inverse of the matrix only if the value is not in cache}
>>>>>>> 259f2c255346c8c138f3532bd81d7700c9ea9538
\usage{
cacheSolve(x, ...)
}
\arguments{
<<<<<<< HEAD
\item{x}{special matrix object with cache capabilities}

\item{...}{additional arguments to the regular \code{solve} function}
}
\description{
\code{cacheSolve} computes the inverse of a matrix
}
\details{
The method intelligently uses cached data to avoid re-calculation
}
\examples{
x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))
cacheSolve(x)
}
\author{
Mario Gamboa
}
\keyword{cache}
=======
\item{x}{the matrix to be "inversed"}
}
\value{
the inversed matrix

Example: cacheSolve(x)
}
\description{
\code{cacheSolve} function that computes the inverse of the matrix only if the value is not in cache
}
>>>>>>> 259f2c255346c8c138f3532bd81d7700c9ea9538

