% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
\title{\code{cacheSolve} function that computes the inverse of the matrix only if the value is not in cache}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{the matrix to be "inversed"}
}
\value{
the inversed matrix

Example: cacheSolve(x)
}
\description{
\code{cacheSolve} function that computes the inverse of the matrix only if the value is not in cache
}

