% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cacheSolve}
\alias{cacheSolve}
\title{Provides a cached varsion of the matrix inversion.}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{a list coming out of makeCacheMatrix}
}
\value{
The inverted matrix referenced by x.
}
\description{
Provides a cached varsion of the matrix inversion.
}
\examples{
amatrix<-makeCacheMatrix(matrix(c(1,2,3,4), nrow=2, ncol=2))
cacheSolve(amatrix)
# first time, gets computed
amatrix$getinverse()
# second time, get cached version
amatrix$getinverse()
}

