% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cachematrix.R
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{A object that contains a matrix and its inverse matrix}
\usage{
makeCacheMatrix(matrix = matrix())
}
\arguments{
\item{matrix}{a matrix object}
}
\value{
a object list containing the methods to access the cache
}
\description{
A object that contains a matrix and its inverse matrix
}
\examples{
m <- makeCacheMatrix(matrix(nrow=3, ncol=3, 3))
}
\author{
Andre B.
}
\seealso{
\code{\link{cacheSolve}} function for cache solving
}

