% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Create a custom matrix with inverse matrix caching capabilities}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{a matrix}
}
\description{
\code{customMatrix} returns a custom matrix object with caching
capabilities
}
\examples{
makeCacheMatrix()
myMatrix <- matrix(rnorm(1e6), 5,5)
myCustomMatrix <- makeCacheMatrix(myMatrix)
myCustomMatrix$getinverse()
}
\keyword{Coursera,}
\keyword{Programming,}
\keyword{ProgrammingAssignment2}
\keyword{R}

