% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachematrix.R
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{A cache matrix}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{the matrix}
}
\value{
A list of function :
\item{set}{set the value of the matrix}
\item{get}{get the value of the matrix}
\item{setInverse}{set the value of the inverse matrix}
\item{getInverse}{get the value of the inverse matrix}
}
\description{
This function `makeCacheMatrix` creates an special 
matrix  which allows to keep in memory the inverse of its matrix
}

