% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Create matrix object with cache capabilities}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{the matrix definition for the object to be created}
}
\description{
\code{makeCacheMatrix} returns a matrix cache object based on its argument
}
\details{
The created matrix object has several functions associated:
\itemize{
   \item set the value of the matrix
   \item get the value of the matrix
   \item set the value of the solved/inversed matrix
   \item get the value of the solved/inversed matrix
}
}
\examples{
x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))
}
\author{
Mario Gamboa
}
\keyword{cache}

