% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{\code{makeCacheMatrix} creates a "matrix" object, with several functions associated
* set the value of the matrix
* get the value of the matrix
* set the value of the solved/inversed matrix
* get the value of the solved/inversed matrix}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{the matrix to which the inverse function will be calculated for}
}
\value{
a list of functions associated with the matrix

Example: `x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))`
}
\description{
\code{makeCacheMatrix} creates a "matrix" object, with several functions associated
* set the value of the matrix
* get the value of the matrix
* set the value of the solved/inversed matrix
* get the value of the solved/inversed matrix
}

