% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
<<<<<<< HEAD
\title{Create matrix object with cache capabilities}
=======
\title{\code{makeCacheMatrix} creates a "matrix" object, with several functions associated
* set the value of the matrix
* get the value of the matrix
* set the value of the solved/inversed matrix
* get the value of the solved/inversed matrix}
>>>>>>> 259f2c255346c8c138f3532bd81d7700c9ea9538
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
<<<<<<< HEAD
\item{x}{the matrix definition for the object to be created}
}
\description{
\code{makeCacheMatrix} returns a matrix cache object based on its argument
}
\details{
The created matrix object has several functions associated:
\itemize{
   \item set the value of the matrix
   \item get the value of the matrix
   \item set the value of the solved/inversed matrix
   \item get the value of the solved/inversed matrix
}
}
\examples{
x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))
}
\author{
Mario Gamboa
}
\keyword{cache}
=======
\item{x}{the matrix to which the inverse function will be calculated for}
}
\value{
a list of functions associated with the matrix

Example: `x <- makeCacheMatrix(rbind(c(1, -1/4), c(-1/4, 1)))`
}
\description{
\code{makeCacheMatrix} creates a "matrix" object, with several functions associated
* set the value of the matrix
* get the value of the matrix
* set the value of the solved/inversed matrix
* get the value of the solved/inversed matrix
}
>>>>>>> 259f2c255346c8c138f3532bd81d7700c9ea9538

