% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Provides support for matrix inversion caching.}
\usage{
makeCacheMatrix(x = matrix())
}
\arguments{
\item{x}{A reference matrix}
}
\value{
The list of closures (set, get, setinverse, getinverse) giving access to the matrix and
  its cached inverse.
}
\description{
Provides support for matrix inversion caching.
}
\examples{
amatrix<-makeCacheMatrix(matrix(c(1,2,3,4), nrow=2, ncol=2))
 amatrix$get()
 amatrix$set(matrix(c(5,6,7,8), nrow=2, ncol=2))
 amatrix$setinverse<-solve(m)
 i<-amatrix$getinverse()
}

