% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeVector}
\alias{makeVector}
\title{Special Vector object with cache}
\usage{
makeVector(x = numeric())
}
\arguments{
\item{x}{vector to which the mean function will be calculated for}
}
\description{
\code{makeVector} Creates a vector object with cache features
}
\details{
creates a special "vector", which is a list containing a functions to:
\itemize{
   \item set the value of the vector
   \item get the value of the vector
   \item set the value of the mean
   \item get the value of the mean
}
}
\examples{
x <- makeVector(c(10,20,5,18))
x$get()
}
\author{
Mario Gamboa
}
\keyword{cache}

